renamed Limelight folder
